test: 新增 react-native-base64 Demo
